- Move interfaces into separate folder - Adjust namespaces
