+++ NDJH.Hypixel.API.DependencyInjection/ServiceExtensions.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 using NDJH.Hypixel.API.Configuration;
 using NDJH.Hypixel.API.Exceptions;
 using NDJH.Hypixel.API.Services;
+using NDJH.Hypixel.API.Services.interfaces;
 
 namespace NDJH.Hypixel.API.DependencyInjection;
