+++ NDJH.Hypixel.API/InputType.cs
@@ -1,0 +1,8 @@
+﻿namespace NDJH.Hypixel.API;
+
+public enum InputType
+{
+    Uuid,
+    Player,
+    Profile
+}
