+++ NDJH.Hypixel.API/Services/Abstractions/IOther.cs
@@ -1,0 +1,9 @@
+namespace NDJH.Hypixel.API.Services.interfaces;
+
+public interface IOther
+{
+    public Task GetBoosters();
+    public Task GetPlayerCounts();
+    public Task GetLeaderboards();
+    public Task GetPunishmentStats();
+}
