--- NDJH.Hypixel.API/Services/InputType.cs
@@ -1,7 +1,0 @@
-﻿namespace NDJH.Hypixel.API.Services;
-
-public enum InputType {
-    Uuid,
-    Player,
-    Profile
-}
